requirements.txt
accelerate==1.6.0
matplotlib==3.10.1
numpy==2.1.3
pandas==2.2.3
pydantic==2.11.2
python-dotenv==1.1.0
python-telegram-bot==22.0
scikit-learn==1.6.1
tensorflow==2.19.0
tf_keras==2.19.0
torch==2.6.0
transformers==4.51.1
